# opt: Improve building and traversing lowest cost expression tree