--- a/pkg/sql/opt/operator.go
+++ b/pkg/sql/opt/operator.go
@@ -21,7 +21,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/types"
 )
 
-//go:generate optgen -out operator.og.go ops ops/scalar.opt ops/relational.opt ops/enforcer.opt
+//go:generate optgen -out operator.og.go ops ops/*.opt
 
 // Operator describes the type of operation that a memo expression performs.
 // Some operators are relational (join, select, project) and others are scalar