--- a/pkg/sql/opt/xform/expr.go
+++ b/pkg/sql/opt/xform/expr.go
@@ -22,18 +22,19 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/util/treeprinter"
 )
 
-//go:generate optgen -out expr.og.go exprs ../ops/scalar.opt ../ops/relational.opt ../ops/enforcer.opt
+//go:generate optgen -out expr.og.go exprs ../ops/*.opt
 
 // ExprView provides a view of a single tree in the memo's forest of query plan
-// trees (see comment in memo.go for more details about the memo forest). Given
-// a particular root memo group and a set of required physical properties,
-// there is one designated lowest cost tree in the memo's forest that provides
-// those properties. ExprView overlays the memo groups with a view of that
-// single tree, exposing methods like ChildCount and Child that enable
-// efficient traversal over that one tree in the forest. ExprView is 24 bytes
-// on a 64-bit machine, and is immutable after construction, so it can be
-// constructed on the stack and passed by value. In addition, it is lazily
-// constructed only when needed (generally when the Child method is called).
+// trees (see comment in memo.go for more details about the memo forest). For
+// a root memo group and a set of required physical properties, there is one
+// designated lowest cost tree (called the "best" expression) in the memo's
+// forest that provides those properties. ExprView overlays the memo groups
+// with a view of that single tree, exposing methods like ChildCount and Child
+// that enable efficient traversal over that one tree in the forest. ExprView
+// is 16 bytes on a 64-bit machine, and is immutable after construction, so it
+// can be constructed on the stack and passed by value. In addition, it is
+// lazily constructed only when needed (generally when the Child method is
+// called).
 //
 // As an example, consider a memo forest with two alternate join orders. There
 // are two trees in this forest:
@@ -58,62 +59,45 @@ import (
 //      +----| scan b |
 //           +--------+
 //
-// Don't reorder fields without checking the size of ExprView stays under 24
-// bytes.
+// Don't reorder fields without checking the impact on the size of ExprView.
 type ExprView struct {
 	// mem references the memo which holds the expression forest.
 	mem *memo
 
-	// loc locates a group within the memo, and a memo expression within that
-	// group.
-	loc memoLoc
+	// group is the identifier of the memo group to which the expression belongs.
+	group opt.GroupID
 
-	// op is the type of the expression (select, inner-join, and, or, etc).
+	// op is the type of the expression.
 	op opt.Operator
 
-	// required are the physical properties that this expression must provide.
-	required opt.PhysicalPropsID
+	// best references the group's lowest cost expression. If best is the
+	// special normBestOrdinal, then this ExprView is traversing the normalized
+	// expression tree rather than the lowest cost tree.
+	best bestOrdinal
 }
 
-// makeExprView creates a new ExprView instance that references the expression
-// in the specified group that satisfies the required properties at the lowest
-// cost. NormPhysPropsID is a special set of required properties that yields a
-// view over the normalized, logical expression tree. This will contain no
-// enforcers and is not necessarily the lowest cost expression.
-//
-// Think of a PhysicalPropsID as selecting one tree out of the memo forest.
-// That tree can consist of all normalized expressions (i.e. first expression
-// in each group), or it can consist of lowest cost expressions (i.e.
-// expressions from bestExprs), or any other criteria we decide is useful.
-//
-// NOTE: While the NormPhysPropsID property set always works, other property
-//       sets will only work once the optimizer has been invoked with that set,
-//       so that it's had an opportunity to compute the lowest cost path.
-func makeExprView(mem *memo, group opt.GroupID, required opt.PhysicalPropsID) ExprView {
-	mgrp := mem.lookupGroup(group)
+// makeExprView creates a new ExprView instance that references the given
+// expression, which is the lowest cost expression in its group for a
+// particular set of physical properties. Note that the group must have already
+// been optimized with respect to that set of properties. Children of this
+// expression will in turn be the lowest cost expressions in their respective
+// groups, and so on.
+func makeExprView(mem *memo, best bestExprID) ExprView {
+	mgrp := mem.lookupGroup(best.group)
+	be := mgrp.bestExpr(best.ordinal)
+	return ExprView{mem: mem, group: best.group, op: be.op, best: best.ordinal}
+}
 
-	if required == opt.NormPhysPropsID {
-		// NormPhysPropsID is the special id used to traverse the expression
-		// tree before it's been fully explored and costed (and bestExprs has
-		// been populated). Create the ExprView over the group's normalized
-		// expression, which is always the first expression in the group.
-		return ExprView{
-			mem:      mem,
-			loc:      makeNormLoc(group),
-			op:       mgrp.lookupExpr(normExprID).op,
-			required: required,
-		}
-	}
-
-	// Find the lowest cost expression in the group that provides the required
-	// properties. That becomes the root of the expression tree.
-	best := mgrp.lookupBestExpr(required)
-	return ExprView{
-		mem:      mem,
-		loc:      memoLoc{group: group, expr: best.lowest},
-		op:       best.op,
-		required: required,
-	}
+// makeNormExprView constructs an ExprView that traverses the normalized
+// logical expression tree, rather than the lowest cost physical tree. The
+// normalized expression tree will contain no enforcer expressions and physical
+// properties are not available. This view is useful when testing or debugging,
+// and to traverse the tree before it's been fully explored and costed (and
+// bestExprs have been populated). See the struct comment in factory.go for
+// more details about the normalized expression tree.
+func makeNormExprView(mem *memo, group opt.GroupID) ExprView {
+	op := mem.lookupNormExpr(group).op
+	return ExprView{mem: mem, group: group, op: op, best: normBestOrdinal}
 }
 
 // Operator returns the type of the expression.
@@ -123,7 +107,7 @@ func (ev ExprView) Operator() opt.Operator {
 
 // Logical returns the set of logical properties that this expression provides.
 func (ev ExprView) Logical() *LogicalProps {
-	return &ev.mem.lookupGroup(ev.loc.group).logical
+	return &ev.mem.lookupGroup(ev.group).logical
 }
 
 // Physical returns the physical properties required of this expression, such
@@ -131,34 +115,55 @@ func (ev ExprView) Logical() *LogicalProps {
 // properties *provided* by this expression, but those *required* of it by its
 // parent expression, or by the ExprView creator.
 func (ev ExprView) Physical() *opt.PhysicalProps {
-	return ev.mem.lookupPhysicalProps(ev.required)
+	if ev.best == normBestOrdinal {
+		panic("physical properties are not available when traversing the normalized tree")
+	}
+	return ev.mem.lookupPhysicalProps(ev.lookupBestExpr().required)
 }
 
 // Group returns the memo group containing this expression.
 func (ev ExprView) Group() opt.GroupID {
-	return ev.loc.group
+	return ev.group
 }
 
 // Child returns the nth expression that is an input to this parent expression.
 // It panics if the requested child does not exist.
 func (ev ExprView) Child(nth int) ExprView {
-	group := ev.ChildGroup(nth)
-
-	if ev.required == opt.NormPhysPropsID {
-		// NormPhysPropsID is the special id used to traverse the expression
-		// tree before it's been fully explored and costed (and bestExprs has
-		// been populated).
-		return makeExprView(ev.mem, group, opt.NormPhysPropsID)
+	if ev.best == normBestOrdinal {
+		// normBestOrdinal is the special BestExpr index that indicates traversal
+		// of the normalized expression tree, regardless of whether it's the
+		// lowest cost.
+		group := ev.ChildGroup(nth)
+		return makeNormExprView(ev.mem, group)
 	}
+	return makeExprView(ev.mem, ev.lookupBestExpr().child(nth))
+}
 
-	required := ev.mem.physPropsFactory.constructChildProps(ev, nth)
-	return makeExprView(ev.mem, group, required)
+// ChildCount returns the number of expressions that are inputs to this
+// parent expression.
+func (ev ExprView) ChildCount() int {
+	if ev.best == normBestOrdinal {
+		return ev.mem.lookupNormExpr(ev.group).childCount()
+	}
+	return ev.lookupBestExpr().childCount()
+}
+
+// ChildGroup returns the memo group containing the nth child of this parent
+// expression.
+func (ev ExprView) ChildGroup(nth int) opt.GroupID {
+	if ev.best == normBestOrdinal {
+		return ev.mem.lookupNormExpr(ev.group).childGroup(ev.mem, nth)
+	}
+	return ev.lookupBestExpr().child(nth).group
 }
 
 // Private returns any private data associated with this expression, or nil if
 // there is none.
 func (ev ExprView) Private() interface{} {
-	return ev.mem.lookupPrivate(ev.privateID())
+	if ev.best == normBestOrdinal {
+		return ev.mem.lookupNormExpr(ev.group).private(ev.mem)
+	}
+	return ev.mem.lookupExpr(ev.lookupBestExpr().eid).private(ev.mem)
 }
 
 // Metadata returns the metadata that's specific to this expression tree. Some
@@ -178,6 +183,10 @@ func (ev ExprView) String() string {
 
 func (ev ExprView) lookupChildGroup(nth int) *memoGroup {
 	return ev.mem.lookupGroup(ev.ChildGroup(nth))
+}
+
+func (ev ExprView) lookupBestExpr() *bestExpr {
+	return ev.mem.lookupGroup(ev.group).bestExpr(ev.best)
 }
 
 func (ev ExprView) format(tp treeprinter.Node) {
@@ -336,89 +345,4 @@ func (ev ExprView) formatPresentation(presentation opt.Presentation, tp treeprin
 		logProps.formatCol(col.Label, col.Index, ev.Metadata(), &buf)
 	}
 	tp.Child(buf.String())
-}
-
-// opLayout describes the "layout" of each op's children. It contains multiple
-// fields:
-//
-//  - baseCount (bits 0,1):
-//      number of children, excluding any list; the children are in
-//      state[0] through state[baseCount-1], unless the op is an enforcer.
-//
-//  - list (bits 2,3):
-//      0 if op has no list, otherwise 1 + position of the list in state
-//      (specifically Offset=state[list-1] Length=state[list]).
-//
-//  - priv (bits 4,5):
-//      0 if op has no private, otherwise 1 + position of the private in state.
-//
-//  - isEnforcer (bit 7): if set, this op is an enforcer; which always has one
-//      child (which is the same group). In this case baseCount is 1 (even
-//      though the child is implicit and not stored in state[0]).
-//
-// The table of values (opLayoutTable) is generated by optgen.
-type opLayout uint8
-
-func (val opLayout) baseCount() uint8 {
-	return uint8(val) & 3
-}
-
-func (val opLayout) list() uint8 {
-	return (uint8(val) >> 2) & 3
-}
-
-func (val opLayout) priv() uint8 {
-	return (uint8(val) >> 4) & 3
-}
-
-func (val opLayout) isEnforcer() bool {
-	return (uint8(val) >> 7) != 0
-}
-
-func makeOpLayout(baseCount, list, priv, isEnforcer uint8) opLayout {
-	return opLayout(baseCount | (list << 2) | (priv << 4) | (isEnforcer << 7))
-}
-
-// ChildCount returns the number of expressions that are inputs to this
-// parent expression.
-func (ev ExprView) ChildCount() int {
-	layout := opLayoutTable[ev.op]
-	baseCount := layout.baseCount()
-	list := layout.list()
-	if list == 0 {
-		return int(baseCount)
-	}
-	e := ev.mem.lookupExpr(ev.loc)
-	return int(baseCount) + int(e.state[list])
-}
-
-// ChildGroup returns the memo group containing the nth child of this parent
-// expression.
-func (ev ExprView) ChildGroup(n int) opt.GroupID {
-	layout := opLayoutTable[ev.op]
-	baseCount := layout.baseCount()
-	e := ev.mem.lookupExpr(ev.loc)
-	if n < int(baseCount) {
-		if layout.isEnforcer() {
-			// Enforcers have a single child which is the same group they're in.
-			return ev.loc.group
-		}
-		return opt.GroupID(e.state[n])
-	}
-	n -= int(baseCount)
-	list := layout.list()
-	if list != 0 && n < int(e.state[list]) {
-		listID := opt.ListID{Offset: e.state[list-1], Length: e.state[list]}
-		return ev.mem.lookupList(listID)[n]
-	}
-	panic("child index out of range")
-}
-
-func (ev ExprView) privateID() opt.PrivateID {
-	priv := opLayoutTable[ev.op].priv()
-	if priv == 0 {
-		return 0
-	}
-	e := ev.mem.lookupExpr(ev.loc)
-	return opt.PrivateID(e.state[priv-1])
 }