… pkg/sql/opt/xform/expr.og.go
… pkg/sql/opt/xform/factory.og.go
… pkg/sql/opt/xform/memo.go
… pkg/sql/opt/xform/memo_expr.go
… pkg/sql/opt/xform/memo_group.go
… pkg/sql/opt/xform/optimizer.go
… pkg/sql/opt/xform/physical_props_factory.go