This PR contains several improvements with how we build and traverse
the lowest cost expression tree (i.e. bestExpr):

1. bestExpr now stores its (lowest cost) children. This improves
   traversal times. More importantly, it will enable us to make more
   complex decisions about what physical properties to require of
   children. For example, the MergeJoin operator will be able to pick
   an arbitrary pair of left/right child expressions to merge.

2. Temporary optimization state is now separate from bestExpr. This
   sets us up to split out memo-related structs into another package.

3. memoGroup and bestExpr optimize for the case where there is only
   one expression in the group, and one bestExpr in the group. This is
   very common, such as in the case of a scalar expression. By storing
   the first expr and the first N best children inline, we avoid
   allocations in the common case.

4. ExprView now use bestExpr to traverse the lowest cost expression
   tree, which is fast because bestExpr now stores its children.

The sum total of these improvements significantly improved the
ExprViewBenchmark:

  name          old time/op   new time/op
  ExprView/0-8  365ns ± 1%    233ns ± 2%

Release note: None