@RaduBerinde 2018-03-16
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>

Nice improvement!  I think this also makes the traversal logic more clear - having to derive the child required properties each time made things complicated.

---

Review status: 0 of 16 files reviewed at latest revision, 1 unresolved discussion, some commit checks pending.

---

*[pkg/sql/opt/xform/memo_group.go, line 113 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/23971#-L7kppysnpFKoZS9onhZ:-L7kppytHSgclzq_TP5I:b-poqbd0) ([raw file](https://github.com/cockroachdb/cockroach/blob/d87990c935046b301ceb4c0ebd4bc340351ca699/pkg/sql/opt/xform/memo_group.go#L113)):*
> ```Go
> // bestExprCount returns the number of bestExprs in the group.
> func (g *memoGroup) bestExprCount() int {
> 	if g.firstBestExpr.required == 0 {
> ```

[nit] This was a bit confusing at first (I incorrectly thought that 0 means "no properties required"). I'd hide the check behind a `bestExpr.isInitialized()` method and use that, then the intent is clear.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/23971#-:-L7kqEVstYEfYJeCltAn:b8upkvh)*
<!-- Sent from Reviewable.io -->