@andy-kimball 2018-03-16
Review status: 0 of 19 files reviewed at latest revision, 1 unresolved discussion.

---

*[pkg/sql/opt/xform/memo_group.go, line 113 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/23971#-L7kppysnpFKoZS9onhZ:-L7l4WM7gcmkYhsC9OBx:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/d87990c935046b301ceb4c0ebd4bc340351ca699/pkg/sql/opt/xform/memo_group.go#L113)):*
<details><summary><i>Previously, RaduBerinde wrote…</i></summary><blockquote>

[nit] This was a bit confusing at first (I incorrectly thought that 0 means "no properties required"). I'd hide the check behind a `bestExpr.isInitialized()` method and use that, then the intent is clear.
</blockquote></details>

Done.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/23971)*
<!-- Sent from Reviewable.io -->