@rytaft 2018-03-19
Review status: 0 of 19 files reviewed at latest revision, 2 unresolved discussions, all commit checks successful.

---

*[pkg/sql/opt/xform/best_expr.go, line 85 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/23971#-L7yWjxJlQiMRXCDI_sJ:-L7yWjxJlQiMRXCDI_sK:bt3myxm) ([raw file](https://github.com/cockroachdb/cockroach/blob/227bfac7761eff07fe9fd01d18583cd3a94e9794/pkg/sql/opt/xform/best_expr.go#L85)):*
> ```Go
> 		return be.initial[nth]
> 	}
> 	return be.remainder[nth-opt.MaxOperands]
> ```

Shouldn't this be `nth-be.initialCount`? What if `initialCount` is less than `opt.MaxOperands`?

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/23971)*
<!-- Sent from Reviewable.io -->