@andy-kimball 2018-03-19
*[pkg/sql/opt/xform/best_expr.go, line 85 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/23971#-L7yWjxJlQiMRXCDI_sJ:-L7yZefFj5JGWYCLsrI4:b-w2f4sq) ([raw file](https://github.com/cockroachdb/cockroach/blob/227bfac7761eff07fe9fd01d18583cd3a94e9794/pkg/sql/opt/xform/best_expr.go#L85)):*
<details><summary><i>Previously, rytaft wrote…</i></summary><blockquote>

Shouldn't this be `nth-be.initialCount`? What if `initialCount` is less than `opt.MaxOperands`?
</blockquote></details>

If it were less than `opt.MaxOperands` then `be.remainder` wouldn't be used. Only once we've "filled" `be.initial` do we start to add to `be.remainder`.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/23971)*
<!-- Sent from Reviewable.io -->