@rytaft 2018-03-19
*[pkg/sql/opt/xform/best_expr.go, line 85 at r2](https://beta.reviewable.io/reviews/cockroachdb/cockroach/23971#-L7yWjxJlQiMRXCDI_sJ:-L7yaK_0CEvAGi_yIaNW:bjtb84p) ([raw file](https://github.com/cockroachdb/cockroach/blob/227bfac7761eff07fe9fd01d18583cd3a94e9794/pkg/sql/opt/xform/best_expr.go#L85)):*
<details><summary><i>Previously, andy-kimball (Andy Kimball) wrote…</i></summary><blockquote>

If it were less than `opt.MaxOperands` then `be.remainder` wouldn't be used. Only once we've "filled" `be.initial` do we start to add to `be.remainder`.
</blockquote></details>

got it - thanks!

---


*Comments from [Reviewable](https://beta.reviewable.io/reviews/cockroachdb/cockroach/23971)*
<!-- Sent from Reviewable.io -->